requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
mysql-connector-python
gunicorn